pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-report-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pick>=2.6.0",
    "blessed>=1.20.0",
    "jira>=3.10.5",
    "python-dotenv>=1.2.2",
    "typer>=0.25.1",
]

[project.scripts]
jira-report-buddy = "main:main"

[tool.setuptools]
py-modules = ["main", "fuzzy_find", "configure", "jira_buddy"]

[tool.pyright]
reportImplicitOverride = true
